Add static versions of setter functions for use by derived classes.
Derived classes need these functions in order to reduce the amount of code
emitted for template functions that do not actually depend on the template
argument type. They allow code to be written externally that does not
depend on the template argument type, and that code can be called in a very
direct way so the template-dependent code is optimized out.