Small work_queue changes described below...
Fix to swap with an empty function pointer in deque instead of moving.
There is no guarantee that the moved from pointer will be released, and
for some things stored on the queue this will be a big deal.

Also, added void * as an alignment term in the pimp union. Just to be on the
safe side.